Project import generated by Copybara.
GitOrigin-RevId: c1506ca4fca61fedd91c47eff82451322552611a
