+++ MODULE.bazel
@@ -16,7 +16,7 @@ module(name = "org_outernetcouncil_nmts")
 
 bazel_dep(
     name = "gazelle",
-    version = "0.38.0",
+    version = "0.39.1",
 )
 
 bazel_dep(
