+++ cmd/nmtscli/html.go
@@ -23,7 +23,7 @@ import (
 	"slices"
 
 	"github.com/urfave/cli/v2"
-	er "outernetcouncil.org/nmts/lib/entity_relationship"
+	er "outernetcouncil.org/nmts/lib/entityrelationship"
 	npb "outernetcouncil.org/nmts/proto"
 )
