+++ cmd/nmtscli/main.go
@@ -22,7 +22,7 @@ import (
 	"os"
 
 	"github.com/urfave/cli/v2"
-	er "outernetcouncil.org/nmts/lib/entity_relationship"
+	er "outernetcouncil.org/nmts/lib/entityrelationship"
 )
 
 const appName = "nmtscli"
