+++ cmd/nmtscli/prolog.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/ichiban/prolog"
 	"github.com/urfave/cli/v2"
-	er "outernetcouncil.org/nmts/lib/entity_relationship"
+	er "outernetcouncil.org/nmts/lib/entityrelationship"
 	npb "outernetcouncil.org/nmts/proto"
 )
