+++ lib/entityrelationship/entity.go
@@ -12,36 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package entity_relationship
+package entityrelationship
 
 import (
-	"fmt"
-	"os"
+	"reflect"
+	"strings"
 
-	"google.golang.org/protobuf/encoding/prototext"
 	npb "outernetcouncil.org/nmts/proto"
 )
 
-func ReadFragmentFiles(fragmentFilenames []string) (*npb.Fragment, error) {
-	g := &npb.Fragment{}
+func EntityKindStringFromProto(e *npb.Entity) string {
+	if e == nil {
+		return ""
+	}
 
-	for _, f := range fragmentFilenames {
-		data, err := os.ReadFile(f)
-		if err != nil {
-			return nil, fmt.Errorf("reading %q: %w", f, err)
-		}
-		subg := &npb.Fragment{}
-		if err := prototext.Unmarshal(data, subg); err != nil {
-			return nil, fmt.Errorf("parsing %q: %w", f, err)
-		}
+	kind := e.GetKind()
+	if kind == nil {
+		return ""
+	}
 
-		for _, e := range subg.GetEntity() {
-			g.Entity = append(g.Entity, e)
-		}
-		for _, r := range subg.GetRelationship() {
-			g.Relationship = append(g.Relationship, r)
+	tag := reflect.TypeOf(kind).Elem().Field(0).Tag.Get("protobuf")
+	for _, element := range strings.Split(tag, ",") {
+		if strings.HasPrefix(element, "name=") {
+			return strings.ToUpper(strings.TrimPrefix(element, "name="))
 		}
 	}
 
-	return g, nil
+	return ""
 }
